일반 평시로그인까지
?

+++ auction-backend/auction-backend/src/main/java/com/auction/dto/BidDto.java
@@ -6,7 +6,7 @@ public class BidDto {
     private Long id;
     private Long auctionId;
     private String bidder;
-    private int bidAmount;
+    private Long bidAmount;
     private LocalDateTime bidTime;
     private LocalDateTime createdAt;
 
@@ -21,8 +21,8 @@ public class BidDto {
     public String getBidder() { return bidder; }
     public void setBidder(String bidder) { this.bidder = bidder; }
 
-    public int getBidAmount() { return bidAmount; }
-    public void setBidAmount(int bidAmount) { this.bidAmount = bidAmount; }
+    public Long getBidAmount() { return bidAmount; }
+    public void setBidAmount(Long bidAmount) { this.bidAmount = bidAmount; }
 
     public LocalDateTime getBidTime() { return bidTime; }
     public void setBidTime(LocalDateTime bidTime) { this.bidTime = bidTime; }
